refactor: BaseEntity 클래스 경로 수정
- 기존: /travelservice/config/BaseEntity
- 수정: /travelservice/global/BaseEntity
